refactor: Update component aliases and paths
