+++ features/website/About.tsx
@@ -1,6 +1,6 @@
 import { ArrowRight } from "lucide-react";
 
-import { Button } from "@/components/ui/button";
+import { Button } from "@/shared/ui/button";
 
 const AboutSection = () => (
   <section className="text-center mb-16 container mx-auto px-4" id="about">
